correçoes

+++ src/main/java/Grupo3pt/iade/ChavesApp/services/SeasonService.java
@@ -1,10 +1,9 @@
 package Grupo3pt.iade.ChavesApp.services;
 
+import Grupo3pt.iade.ChavesApp.models.Season;
 import Grupo3pt.iade.ChavesApp.repositories.SeasonRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import com.mysql.cj.Session;
 
 import java.util.List;
 import java.util.Optional;
@@ -12,28 +11,26 @@ import java.util.Optional;
 @Service
 public class SeasonService {
 
-    @SuppressWarnings("rawtypes")
+
     @Autowired
     private SeasonRepository seasonRepository;
 
-    @SuppressWarnings("unchecked")
+
     public List<Season> getAllSeasons() {
         return seasonRepository.findAll();
     }
 
-    @SuppressWarnings("unchecked")
     public Optional<Season> getSeasonById(Long id) {
         return seasonRepository.findById(id);
     }
 
-    @SuppressWarnings("unchecked")
-    public Grupo3pt.iade.ChavesApp.controllers.Season createSeason(Season season) {
-        return (Grupo3pt.iade.ChavesApp.controllers.Season) seasonRepository.save(season);
+    public Season createSeason(Season season) {
+        return seasonRepository.save(season);
     }
 
-    @SuppressWarnings("unchecked")
-    public Grupo3pt.iade.ChavesApp.controllers.Season updateSeason(Long id, Season seasonDetails) {
-        return (Grupo3pt.iade.ChavesApp.controllers.Season) seasonRepository.findById(id)
+    
+    public Season updateSeason(Long id, Season seasonDetails) {
+        return seasonRepository.findById(id)
             .map(season -> {
                 season.setName(seasonDetails.getName());
                 season.setYear(seasonDetails.getYear());
@@ -42,8 +39,12 @@ public class SeasonService {
             .orElseThrow(() -> new RuntimeException("Season not found with id " + id));
     }
 
-    @SuppressWarnings("unchecked")
-    public Grupo3pt.iade.ChavesApp.controllers.Season deleteSeason(Long id) {
-        seasonRepository.deleteById(id);
+    
+    public void deleteSeason(Long id) {
+        if (seasonRepository.existsById(id)) {
+            seasonRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Season not found with id " + id);
+        }
     }
 }
